Seperate private shape and index strategies from NumPy extra
