+++ hypothesis-python/src/hypothesis/extra/numpy.py
@@ -19,8 +19,15 @@ from typing import Any, Mapping, NamedTuple, Optional, Sequence, Tuple, Union
 
 import numpy as np
 
-from hypothesis import assume, strategies as st
+from hypothesis import strategies as st
 from hypothesis.errors import InvalidArgument
+from hypothesis.extra.__array_helpers import (
+    BasicIndex,
+    BasicIndexStrategy,
+    BroadcastableShapes,
+    MutuallyBroadcastableShapesStrategy,
+    Shape,
+)
 from hypothesis.internal.conjecture import utils as cu
 from hypothesis.internal.coverage import check_function
 from hypothesis.internal.reflection import proxies
@@ -29,15 +36,7 @@ from hypothesis.strategies._internal.strategies import T, check_strategy
 from hypothesis.strategies._internal.utils import defines_strategy
 from hypothesis.utils.conventions import UniqueIdentifier, not_set
 
-Shape = Tuple[int, ...]
-# flake8 and mypy disagree about `ellipsis` (the type of `...`), and hence:
-BasicIndex = Tuple[Union[int, slice, "ellipsis", np.newaxis], ...]  # noqa: F821
 TIME_RESOLUTIONS = tuple("Y  M  D  h  m  s  ms  us  ns  ps  fs  as".split())
-
-
-class BroadcastableShapes(NamedTuple):
-    input_shapes: Tuple[Shape, ...]
-    result_shape: Shape
 
 
 @defines_strategy(force_reusable_values=True)
@@ -908,147 +907,6 @@ def broadcastable_shapes(
     ).map(lambda x: x.input_shapes[0])
 
 
-class MutuallyBroadcastableShapesStrategy(st.SearchStrategy):
-    def __init__(
-        self,
-        num_shapes,
-        signature=None,
-        base_shape=(),
-        min_dims=0,
-        max_dims=None,
-        min_side=1,
-        max_side=None,
-    ):
-        assert 0 <= min_side <= max_side
-        assert 0 <= min_dims <= max_dims <= 32
-        st.SearchStrategy.__init__(self)
-        self.base_shape = base_shape
-        self.side_strat = st.integers(min_side, max_side)
-        self.num_shapes = num_shapes
-        self.signature = signature
-        self.min_dims = min_dims
-        self.max_dims = max_dims
-        self.min_side = min_side
-        self.max_side = max_side
-
-        self.size_one_allowed = self.min_side <= 1 <= self.max_side
-
-    def do_draw(self, data):
-        # We don't usually have a gufunc signature; do the common case first & fast.
-        if self.signature is None:
-            return self._draw_loop_dimensions(data)
-
-        # When we *do*, draw the core dims, then draw loop dims, and finally combine.
-        core_in, core_res = self._draw_core_dimensions(data)
-
-        # If some core shape has omitted optional dimensions, it's an error to add
-        # loop dimensions to it.  We never omit core dims if min_dims >= 1.
-        # This ensures that we respect Numpy's gufunc broadcasting semantics and user
-        # constraints without needing to check whether the loop dims will be
-        # interpreted as an invalid substitute for the omitted core dims.
-        # We may implement this check later!
-        use = [None not in shp for shp in core_in]
-        loop_in, loop_res = self._draw_loop_dimensions(data, use=use)
-
-        def add_shape(loop, core):
-            return tuple(x for x in (loop + core)[-32:] if x is not None)
-
-        return BroadcastableShapes(
-            input_shapes=tuple(add_shape(l_in, c) for l_in, c in zip(loop_in, core_in)),
-            result_shape=add_shape(loop_res, core_res),
-        )
-
-    def _draw_core_dimensions(self, data):
-        # Draw gufunc core dimensions, with None standing for optional dimensions
-        # that will not be present in the final shape.  We track omitted dims so
-        # that we can do an accurate per-shape length cap.
-        dims = {}
-        shapes = []
-        for shape in self.signature.input_shapes + (self.signature.result_shape,):
-            shapes.append([])
-            for name in shape:
-                if name.isdigit():
-                    shapes[-1].append(int(name))
-                    continue
-                if name not in dims:
-                    dim = name.strip("?")
-                    dims[dim] = data.draw(self.side_strat)
-                    if self.min_dims == 0 and not data.draw_bits(3):
-                        dims[dim + "?"] = None
-                    else:
-                        dims[dim + "?"] = dims[dim]
-                shapes[-1].append(dims[name])
-        return tuple(tuple(s) for s in shapes[:-1]), tuple(shapes[-1])
-
-    def _draw_loop_dimensions(self, data, use=None):
-        # All shapes are handled in column-major order; i.e. they are reversed
-        base_shape = self.base_shape[::-1]
-        result_shape = list(base_shape)
-        shapes = [[] for _ in range(self.num_shapes)]
-        if use is None:
-            use = [True for _ in range(self.num_shapes)]
-        else:
-            assert len(use) == self.num_shapes
-            assert all(isinstance(x, bool) for x in use)
-
-        for dim_count in range(1, self.max_dims + 1):
-            dim = dim_count - 1
-
-            # We begin by drawing a valid dimension-size for the given
-            # dimension. This restricts the variability across the shapes
-            # at this dimension such that they can only choose between
-            # this size and a singleton dimension.
-            if len(base_shape) < dim_count or base_shape[dim] == 1:
-                # dim is unrestricted by the base-shape: shrink to min_side
-                dim_side = data.draw(self.side_strat)
-            elif base_shape[dim] <= self.max_side:
-                # dim is aligned with non-singleton base-dim
-                dim_side = base_shape[dim]
-            else:
-                # only a singleton is valid in alignment with the base-dim
-                dim_side = 1
-
-            for shape_id, shape in enumerate(shapes):
-                # Populating this dimension-size for each shape, either
-                # the drawn size is used or, if permitted, a singleton
-                # dimension.
-                if dim_count <= len(base_shape) and self.size_one_allowed:
-                    # aligned: shrink towards size 1
-                    side = data.draw(st.sampled_from([1, dim_side]))
-                else:
-                    side = dim_side
-
-                # Use a trick where where a biased coin is queried to see
-                # if the given shape-tuple will continue to be grown. All
-                # of the relevant draws will still be made for the given
-                # shape-tuple even if it is no longer being added to.
-                # This helps to ensure more stable shrinking behavior.
-                if self.min_dims < dim_count:
-                    use[shape_id] &= cu.biased_coin(
-                        data, 1 - 1 / (1 + self.max_dims - dim)
-                    )
-
-                if use[shape_id]:
-                    shape.append(side)
-                    if len(result_shape) < len(shape):
-                        result_shape.append(shape[-1])
-                    elif shape[-1] != 1 and result_shape[dim] == 1:
-                        result_shape[dim] = shape[-1]
-            if not any(use):
-                break
-
-        result_shape = result_shape[: max(map(len, [self.base_shape] + shapes))]
-
-        assert len(shapes) == self.num_shapes
-        assert all(self.min_dims <= len(s) <= self.max_dims for s in shapes)
-        assert all(self.min_side <= s <= self.max_side for side in shapes for s in side)
-
-        return BroadcastableShapes(
-            input_shapes=tuple(tuple(reversed(shape)) for shape in shapes),
-            result_shape=tuple(reversed(result_shape)),
-        )
-
-
 # See https://numpy.org/doc/stable/reference/c-api/generalized-ufuncs.html
 # Implementation based on numpy.lib.function_base._parse_gufunc_signature
 # with minor upgrades to handle numeric and optional dimensions.  Examples:
@@ -1308,58 +1166,6 @@ def mutually_broadcastable_shapes(
     )
 
 
-class BasicIndexStrategy(st.SearchStrategy):
-    def __init__(self, shape, min_dims, max_dims, allow_ellipsis, allow_newaxis):
-        assert 0 <= min_dims <= max_dims <= 32
-        st.SearchStrategy.__init__(self)
-        self.shape = shape
-        self.min_dims = min_dims
-        self.max_dims = max_dims
-        self.allow_ellipsis = allow_ellipsis
-        self.allow_newaxis = allow_newaxis
-
-    def do_draw(self, data):
-        # General plan: determine the actual selection up front with a straightforward
-        # approach that shrinks well, then complicate it by inserting other things.
-        result = []
-        for dim_size in self.shape:
-            if dim_size == 0:
-                result.append(slice(None))
-                continue
-            strategy = st.integers(-dim_size, dim_size - 1) | st.slices(dim_size)
-            result.append(data.draw(strategy))
-        # Insert some number of new size-one dimensions if allowed
-        result_dims = sum(isinstance(idx, slice) for idx in result)
-        while (
-            self.allow_newaxis
-            and result_dims < self.max_dims
-            and (result_dims < self.min_dims or data.draw(st.booleans()))
-        ):
-            result.insert(data.draw(st.integers(0, len(result))), np.newaxis)
-            result_dims += 1
-        # Check that we'll have the right number of dimensions; reject if not.
-        # It's easy to do this by construction iff you don't care about shrinking,
-        # which is really important for array shapes.  So we filter instead.
-        assume(self.min_dims <= result_dims <= self.max_dims)
-        # This is a quick-and-dirty way to insert ..., xor shorten the indexer,
-        # but it means we don't have to do any structural analysis.
-        if self.allow_ellipsis and data.draw(st.booleans()):
-            # Choose an index; then replace all adjacent whole-dimension slices.
-            i = j = data.draw(st.integers(0, len(result)))
-            while i > 0 and result[i - 1] == slice(None):
-                i -= 1
-            while j < len(result) and result[j] == slice(None):
-                j += 1
-            result[i:j] = [Ellipsis]
-        else:
-            while result[-1:] == [slice(None, None)] and data.draw(st.integers(0, 7)):
-                result.pop()
-        if len(result) == 1 and data.draw(st.booleans()):
-            # Sometimes generate bare element equivalent to a length-one tuple
-            return result[0]
-        return tuple(result)
-
-
 @defines_strategy()
 def basic_indices(
     shape: Shape,
@@ -1420,7 +1226,7 @@ def basic_indices(
         min_dims=min_dims,
         max_dims=max_dims,
         allow_ellipsis=allow_ellipsis,
-        allow_newaxis=allow_newaxis,
+        allow_none=allow_newaxis,
     )
